Added ability to use CM values in height.
Added and modified a few lines of your code. Now we can send '75cm' and have the desk move to that height. Also added functions to calculate from cmToBufferNumber and BufferNumberToCm. Also changed the user input value to uppercase and removed the OR conditions to check for uppercase. So now we can set the target value with a more understandable numbers.